Adding BigtableSession.
BigtableSession will contain all of the Channel creation logic from BigtableChannels, the creation of Executors from BigtableOptionsFactory and the close logic from AbstractBigtableConnection.

Refactoring grpc-interface project to move code into BigtableSession
Moving custom event loop, retryExecutor and TransportOption creation from BigtableOptions and BigtableChannels to BigtableSession
Moving all uses of BigtableChannels.createChannel and *GrpcClient creation to BigtableSession
Moved all logic related to closing Channels and ExecutorServices into BigtableSession.

Moving the default port from BigtableOptionsFactory to BigtableOptions.

Adding comments, moving artifacts to more appropriate locations.
