+++ bigtable-grpc-interface/src/test/java/com/google/cloud/bigtable/config/BigtableOptionsTestFactory.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2014 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.bigtable.config;
+
+/**
+ * This class provides a back door approach to creating BigtableOptions that are not fully
+ * configured.  This bypasses the BigtableOptions Precondition checks.
+ */
+public class BigtableOptionsTestFactory {
+
+  public static BigtableOptions build() {
+    return new BigtableOptions();
+  }
+}
