+++ bigtable-grpc-interface/src/test/java/com/google/cloud/bigtable/grpc/BigtableSessionTests.java
@@ -24,6 +24,8 @@ import com.google.bigtable.v1.CheckAndMutateRowRequest;
 import com.google.bigtable.v1.MutateRowRequest;
 import com.google.bigtable.v1.Mutation;
 import com.google.bigtable.v1.Mutation.SetCell;
+import com.google.cloud.bigtable.config.BigtableOptions;
+import com.google.cloud.bigtable.config.BigtableOptionsTestFactory;
 import com.google.common.base.Predicate;
 
 import org.junit.Test;
@@ -35,12 +37,14 @@ import io.grpc.MethodDescriptor;
 import java.util.Map;
 
 @RunWith(JUnit4.class)
-public class BigtableChannelsTest {
+public class BigtableSessionTests {
 
-  @SuppressWarnings("unchecked")
+  @SuppressWarnings({ "unchecked", "resource" })
   @Test
-  public void createMethodRetryMap() {
-    Map<MethodDescriptor<?, ?>, Predicate<?>> map = BigtableChannels.createMethodRetryMap();
+  public void createMethodRetryMap() throws Exception {
+    BigtableOptions options = BigtableOptionsTestFactory.build();
+    BigtableSession bs = new BigtableSession(options, null, null, null);
+    Map<MethodDescriptor<?, ?>, Predicate<?>> map = bs.createMethodRetryMap();
     for (MethodDescriptor<?, ?> method: BigtableServiceGrpc.CONFIG.methods()) {
       if (method == BigtableServiceGrpc.CONFIG.mutateRow) {
         assertMutateRowPredicate((Predicate<MutateRowRequest>) map.get(method));
