+++ bigtable-hbase/src/test/java/com/google/cloud/bigtable/hbase/TestClusterAPI.java
@@ -53,7 +53,7 @@ import com.google.bigtable.admin.cluster.v1.ListZonesResponse;
 import com.google.bigtable.admin.cluster.v1.Zone;
 import com.google.cloud.bigtable.config.BigtableOptions;
 import com.google.cloud.bigtable.grpc.BigtableClusterAdminClient;
-import com.google.cloud.bigtable.grpc.BigtableClusterAdminGrpcClient;
+import com.google.cloud.bigtable.grpc.BigtableSession;
 import com.google.common.util.concurrent.UncheckedExecutionException;
 import com.google.longrunning.GetOperationRequest;
 import com.google.longrunning.Operation;
@@ -81,7 +81,8 @@ public class TestClusterAPI {
     }
 
     BigtableOptions originalOptions = BigtableOptionsFactory.fromConfiguration(configuration);
-    BigtableClusterAdminClient client = createClusterAdminStub(originalOptions);
+    BigtableSession originalSession = new BigtableSession(originalOptions, Executors.newFixedThreadPool(10));
+    BigtableClusterAdminClient client = originalSession.getClusterAdminClient();
 
     String projectId = originalOptions.getProjectId();
     List<Cluster> clusters = getClusters(client, projectId);
@@ -169,13 +170,6 @@ public class TestClusterAPI {
     }
     throw new IllegalStateException(String.format(
       "Waited %d seconds and operation was not complete", maxSeconds));
-  }
-
-  private BigtableClusterAdminClient createClusterAdminStub(BigtableOptions bigtableOptions)
-      throws IOException {
-    return BigtableClusterAdminGrpcClient.createClient(
-      bigtableOptions.getClusterAdminTransportOptions(), bigtableOptions.getChannelOptions(),
-      Executors.newFixedThreadPool(10));
   }
 
   private List<Zone> getZones(BigtableClusterAdminClient client, String projectId) {
